requirements.txt
langchain-openai
langgraph
faiss-cpu
python-dotenv
langchain-community